// packages/integration/generated-matrix/models.ts
export interface MxBase1 {
  field1: string
}

export interface MxBase2 {
  field2: number
}
// ...
export interface MxTreeNode {
  value: string
  left?: MxTreeNode
  right?: MxTreeNode
}

export interface MxMutualA {
  aName: string
  child?: MxMutualB
}

export interface MxMutualB {
  bName: string
  parent?: MxMutualA
}

export type MxCycleCompA = MxBase1 & {
  next?: MxCycleCompB
}

export type MxCycleCompB = MxCycleCompA | MxBase2